\name{adipocyte}
\docType{data}
\alias{adipocyte}
\title{Mouse adipocyte differentiation gene expression (microarray) data}
\description{
The data object contains a time-course gene expression profiling of the mouse adipocytes through differentiation. For details please refer to the article: 

Ma X, Yang P, Kaplan WH, Lee BH, Wu LE, Yang JY, Yasunaga M, Sato K, Chisholm DJ, James DE. ISL1 regulates peroxisome proliferator-activated receptor gamma activation and early adipogenesis via bone morphogenetic protein 4-dependent and-independent mechanisms. Molecular and Cellular Biology. 2014 Oct 1;34(19):3607-17.
}
