% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTest.R
\name{enrichmentTest}
\alias{enrichmentTest}
\title{Fisher's exact test-based enrichment test}
\usage{
enrichmentTest(clust, annotation, universe, alter = "greater")
}
\arguments{
\item{clust}{a vector of names representing phosphorylation sites that are partitioned in the same cluster}

\item{annotation}{a list with names correspond to kinases and elements correspond to substrates belong to each kinase}

\item{universe}{the universe of names to compare against}

\item{alter}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less"}
}
\value{
a matrix that contains enrichment of each kinase based on the input name vector.
}
\description{
Takes a vector of names representing phosphorylation sites that are partitioned in the same cluster 
and an kinase-substrate annotation. Test for enrichment of the kinase based on the name vector.
}
