\name{hES}
\docType{data}
\alias{hES}
\title{Human embryonic stem cell phosphoproteomics data}
\description{
The data object contains a time-course phosphoprotoemics profiling of the human embryonic stem cells through differentiation. The differentiation
is induced by using PMA and the time points of 30m, 1hr, 6hrs, and 24hrs are measured and the ratio are taken against 0m. For details please refer to
the article: Rigbolt et al. Sci Signal. 4(164):rs3, 2011
}
