\name{PhosphoSite.human}
\docType{data}
\alias{PhosphoSite.human}
\title{PhosphoSitePlus annotations for human}
\description{
The data object contains the annotations of kinases and their conrresponding substrates as phosphorylation sites in human. It is extracted
from the PhosphoSitePlus database. For details of PhosphoSitePlus, please refer to the article: Hornbeck et al. Nucleic Acids Res. 40:D261-70, 2012
}
