\name{PhosphoSite.mouse}
\docType{data}
\alias{PhosphoSite.mouse}
\title{PhosphoSitePlus annotations for mouse}
\description{
The data object contains the annotations of kinases and their conrresponding substrates as phosphorylation sites in mouse. It is extracted
from the PhosphoSitePlus database. For details of PhosphoSitePlus, please refer to the article: Hornbeck et al. Nucleic Acids Res. 40:D261-70, 2012
}
